06-Cut-Out-Code & 07-UMAP-Cluster-Inspection
Cut-Out-Code is a script which just contains all the code excised from Deepti's original pipeline, some of which I ended up using. I figured it could be useful to store this for bookkeeping purposes. 07-UMAP-Cluster-Inspection is used to assess cellType identities of various UMAP clusters from the organoid Seurat Object.